Structure files for ruby gem Geonames
