+++ Gemfile
@@ -1,0 +1,4 @@
+source 'https://rubygems.org'
+
+# Specify gem's dependencies in geonames.gemspec
+gemspec
