+++ README.md
@@ -8,3 +8,12 @@ Geonames files can be found at: http://download.geonames.org/export/dump/
 Used files: 
   cities1000.txt
   countryInfo.txt
+
+### How use
+
+Open gem folder and try with irb or pry
+
+```
+irb -Ilib -rgeonames
+Geonames.parser_cities
+```
